Temporarilu adds a quick acess delete spam button on search results.
Until we have other measures for handling spam, this will add (for sysadmins)
a quick access delete button to each item in the search results.  This should make it
a lot quicker to delete spam.

I expect it is of limited value once the spam handling is done properly, but having
quick-access buttons on the search results seems generally useful.